backend/requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.4,<5.5
django-filter>=25.0,<26.0
django-cors-headers>=4.0,<5.0
pillow>=11.0,<12.0
psycopg[binary]>=3.2,<3.3
